Nodes - Added pin value visualization
Nodes - Update values in UI in realtime
Integer node - Fixed loading by removing some trial & error code

+++ src/Artemis.VisualScripting/Nodes/CustomViewModels/StaticIntegerValueNodeCustomViewModel.cs
@@ -13,7 +13,7 @@ namespace Artemis.VisualScripting.Nodes.CustomViewModels
 
         public int Input
         {
-            get => (int)(long) _node.Storage;
+            get => (int) _node.Storage;
             set
             {
                 _node.Storage = value;
